+++ dockeri2c/docker-compose/eijent_v2/Makefile
@@ -9,7 +9,7 @@ BUILD_ADDR_ARM = tcp://buildkit-arm.bookserver.home:1235 #arm
 BUILD_OPTION = "type=image,push=true,registry.insecure=true"
 
 TARGET = ./Dockerfile
-TARGET_FILE = ./golang-v2
+TARGET_FILE = ./golang
 GO_VERSION = 1.18.2
 DOCKERFILE_FOLDER = .
