--- dockeri2c/docker-compose/eijent_v2/golang-v2/go.mod
@@ -1,18 +1,0 @@
-module app
-
-go 1.18
-
-require (
-	github.com/caarlos0/env/v6 v6.10.1
-	github.com/d2r2/go-dht v0.0.0-20200119175940-4ba96621a218
-	github.com/davecheney/i2c v0.0.0-20140823063045-caf08501bef2
-	github.com/tarm/serial v0.0.0-20180830185346-98f6abe2eb07
-	golang.org/x/sync v0.1.0
-)
-
-require (
-	github.com/d2r2/go-logger v0.0.0-20210606094344-60e9d1233e22 // indirect
-	github.com/d2r2/go-shell v0.0.0-20211022052110-f591c27e3e2e // indirect
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	golang.org/x/sys v0.6.0 // indirect
-)
