--- dockeri2c/docker-compose/eijent_v2/golang/route.go
@@ -1,14 +1,0 @@
-package main
-
-import (
-	"app/webserver"
-)
-
-type Htmldata struct{}
-
-func setupRoute() (Htmldata, error) {
-	output := Htmldata{}
-	return output, nil
-}
-func (t *Htmldata) setupRoute(cfg *webserver.SetupServer) {
-}
