+++ dockeri2c/docker-compose/eijent_v2/golang/webserver/webserver.go
@@ -2,6 +2,7 @@ package webserver
 
 import (
 	"context"
+	"fmt"
 	"log"
 	"net"
 	"net/http"
@@ -16,7 +17,7 @@ import (
 type SetupServer struct {
 	Protocol string `env:"PROTOCOL" envDefault:"tcp"`
 	Hostname string `env:"WEB_HOST" envDefault:""`
-	Port     string `env:"WEB_PORT" envDefault:"8080"`
+	Port     string `env:"WEB_PORT" envDefault:"9140"`
 
 	mux *http.ServeMux
 }
@@ -40,7 +41,7 @@ func NewSetup() (*SetupServer, error) {
 }
 
 func (t *SetupServer) NewServer() (*Server, error) {
-	log.Println("Setupserver", t.Protocol, t.Hostname+":"+t.Port)
+	fmt.Println("Setupserver", t.Protocol, t.Hostname+":"+t.Port)
 	l, err := net.Listen(t.Protocol, t.Hostname+":"+t.Port)
 	if err != nil {
 		return nil, err
@@ -62,7 +63,7 @@ func (s *Server) Run(ctx context.Context) error {
 	defer stop()
 	eg, ctx := errgroup.WithContext(ctx)
 	eg.Go(func() error {
-		log.Println("Start Server")
+		fmt.Println("Start Server")
 		if err := s.srv.Serve((s.l)); err != nil {
 			return err
 		}
@@ -72,6 +73,6 @@ func (s *Server) Run(ctx context.Context) error {
 	if err := s.srv.Shutdown(context.Background()); err != nil {
 		log.Println(err)
 	}
-	log.Println("shutdown")
+	fmt.Println("shutdown")
 	return eg.Wait()
 }
