+++ dockeri2c/docker-compose/eijent_v2/readme.md
@@ -5,6 +5,9 @@ curl localhost:9140/reset -X POST
 ```
 
 ## 履歴
+* ver 0.13
+  * センサーの効率化
+  * CO2センサーの異常時の排除設定
 * ver 0.12
   * 待ち時間が発生するので修正の必要あり
   * DHTのセンサーによる待ち時間がでかいので修正する必要あり
